chore: break out game object component from cell
