+++ packages/frontend/src/components/gameObject/index.ts
@@ -1,0 +1,1 @@
+export * from './GameObject';
