+++ packages/frontend/src/index.css
@@ -86,15 +86,6 @@
   animation-duration: 8000ms;
   animation-iteration-count: infinite;
   animation-timing-function: linear;
-  /* transform: rotate(3deg); */
-  /* transform: rotate(0.3rad);/ */
-  /* transform: rotate(3grad); */
-  /* transform: rotate(.03turn);  */
-}
-
-.diamond {
-  fill: url(#radialGradient-1);
-  position: relative;
 }
 
 .sparkles {
